apps/sample-deep-agent/pyproject.toml
[project]
name = "sample-deep-agent"
version = "0.0.1"
description = "Sample Deep Agent using deepagents framework with research capabilities"
authors = [
    { name = "Haili Zhang", email = "haili.zhang@outlook.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "deepagents>=0.1.3",
    "langgraph-up-devkits",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sample_deep_agent"]

[tool.setuptools.package-dir]
"sample_deep_agent" = "src/sample_deep_agent"

[tool.uv.sources]
langgraph-up-devkits = { workspace = true }

[dependency-groups]
dev = [
    "ruff>=0.13.0",
    "mypy>=1.11.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-dotenv>=0.5.2",
]
cli = [
    "langgraph-cli[inmem]>=0.4.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = ["-v", "--tb=short"]
env_files = [".env"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
